refactor(direct-control): extract unit target action
Refs: openspec/changes/civ7-support-direct-control-modularization

Move only getCiv7UnitTargetAction / requestCiv7UnitTargetAction orchestration, the readUnitTargetAction embedded source, command builder, and bounded post-send stabilizer into packages/civ7-direct-control/src/play/operations/unit-target-action.ts while keeping public facade exports in packages/civ7-direct-control/src/index.ts.

Preserve approval-first send behavior, read/send split, parser label Civ7 unit target action, official target-selection order, bounded post-send no-repeat-after-unverified polling, and re-read HUD/ready-unit guidance.

Verification:
- bun run --cwd packages/civ7-direct-control test -- unit-target-action.test.ts
- bun run --cwd packages/civ7-direct-control build
- bun run --cwd packages/cli test -- game.play.unit-target.test.ts
- bun run test:cli:play
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- git diff --check
- bun run check:cli
- ownership/no-cycle rg scans for unit-target action ownership and facade type-only imports

Local package/source relocation proof only; no runtime or live-game proof claimed.
