+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -39,6 +39,13 @@ hierarchy or oRPC work depends on it:
 | Root inspection and runtime API snapshots | Root inspection types around 829-840, App UI/Tuner snapshots around 199-288 and wrappers around 2097-2349                                                                                                                 | `src/runtime/{inspection,app-ui-snapshot,tuner-health}.ts`                                              | `test/runtime-and-catalog.test.ts` owns App UI/Tuner command routing, App UI snapshots, Tuner readiness, runtime API inspection, playable status, shell classification, unready Tuner handling, and targeted GameInfo row coverage.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               | Preserve shell/playable classification and keep live runtime claims separate from local fake-tuner proof.                                                                                                                                                                                          | runtime proof for runtime-state behavior changes                                | Package proof completed; source extraction still pending.                                                                |
 | Capability catalog and proof/log support  | Static entries around 10089, catalog exports 3594/3616, file/log proof helpers 3694-3758                                                                                                                                  | `src/catalog/*` and `src/proof/*`                                                                       | `test/runtime-and-catalog.test.ts` owns static/runtime/official-resource catalog evidence, ordered proof-log markers, and log rewrite edge cases.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     | Preserve proof-class separation and do not use generated output as source authority.                                                                                                                                                                                                               | runtime proof only for live proof behavior claims                               | Package proof completed; source extraction still pending.                                                                |
 
+Operation row update: `src/play/operations/unit-target-action.ts` now owns
+`getCiv7UnitTargetAction` / `requestCiv7UnitTargetAction` orchestration, the
+embedded `readUnitTargetAction` source, command builder, and bounded post-send
+stabilizer while public facade exports stay in `index.ts`. This completes the
+unit-target source/wrapper relocation; broader generic operation and
+production-choice wrapper composition still need named owner/proof boundaries.
+
 ## Forbidden Owners
 
 - CLI must not own raw socket framing, state discovery, reconnect polling,
